refactor: Separating out all classes into their own files.
