+++ include/moth_ui/animation_track.h
@@ -1,102 +1,13 @@
 #pragma once
 
-#include "moth_ui/utils/interp.h"
+#include "moth_ui/ui_fwd.h"
 
 #include <nlohmann/json_fwd.hpp>
 
-#include <variant>
-#include <string>
 #include <array>
 #include <functional>
 
-namespace nlohmann {
-    template <>
-    struct adl_serializer<std::variant<float, std::string>> {
-        static void from_json(json const& j, std::variant<float, std::string>& var) {
-            size_t const index = j["index"];
-            if (index == 0) {
-                var = static_cast<float>(j["value"]);
-            } else if (index == 1) {
-                var = static_cast<std::string>(j["value"]);
-            }
-        }
-
-        static void to_json(json& j, std::variant<float, std::string> const& var) {
-            j["index"] = var.index();
-            std::visit([&](auto&& value) {
-                j["value"] = value;
-            },
-                       var);
-        }
-    };
-}
-
 namespace moth_ui {
-    struct AnimationClip;
-
-    using KeyframeValue = float;
-
-    class Keyframe {
-    public:
-        Keyframe() = default;
-        Keyframe(int frame, KeyframeValue value)
-            : m_frame(frame)
-            , m_value(value) {}
-
-        int m_frame = 0;
-        KeyframeValue m_value = 0;
-        InterpType m_interpType = InterpType::Linear;
-
-        friend void to_json(nlohmann::json& j, Keyframe const& keyframe) {
-            j["frame"] = keyframe.m_frame;
-            j["value"] = keyframe.m_value;
-            j["interp"] = keyframe.m_interpType;
-        }
-
-        friend void from_json(nlohmann::json const& j, Keyframe& keyframe) {
-            nlohmann::json valueJson = j.at("value");
-            if (valueJson.is_number_float()) {
-                valueJson.get_to(keyframe.m_value);
-            } else {
-                std::variant<float, std::string> oldValueType;
-                valueJson.get_to(oldValueType);
-                keyframe.m_value = std::get<float>(oldValueType);
-            }
-
-            j.at("frame").get_to(keyframe.m_frame);
-            j.at("interp").get_to(keyframe.m_interpType);
-        }
-    };
-
-    class AnimationEvent {
-    public:
-        AnimationEvent() = default;
-        AnimationEvent(int frame, std::string const& name)
-            : m_frame(frame)
-            , m_name(name) {}
-
-        int m_frame = 0;
-        std::string m_name;
-
-        bool operator==(AnimationEvent const& other) const {
-            return m_frame == other.m_frame && m_name == other.m_name;
-        }
-
-        bool operator!=(AnimationEvent const& other) const {
-            return !(*this == other);
-        }
-
-        friend void to_json(nlohmann::json& j, AnimationEvent const& event) {
-            j["frame"] = event.m_frame;
-            j["name"] = event.m_name;
-        }
-
-        friend void from_json(nlohmann::json const& j, AnimationEvent& event) {
-            j.at("frame").get_to(event.m_frame);
-            j.at("name").get_to(event.m_name);
-        }
-    };
-
     class AnimationTrack {
     public:
         enum class Target {
@@ -149,22 +60,8 @@ namespace moth_ui {
 
         void SortKeyframes();
 
-        friend void to_json(nlohmann::json& j, AnimationTrack const& track) {
-            j["target"] = track.m_target;
-            std::vector<Keyframe> keyframes;
-            for (auto& keyframe : track.m_keyframes) {
-                keyframes.push_back(*keyframe);
-            }
-            j["keyframes"] = keyframes;
-        }
-
-        friend void from_json(nlohmann::json const& j, AnimationTrack& track) {
-            track.m_target = j.value("target", Target::Unknown);
-            auto keyframes = j.value("keyframes", std::vector<Keyframe>{});
-            for (auto& keyframe : keyframes) {
-                track.m_keyframes.push_back(std::make_unique<Keyframe>(keyframe));
-            }
-        }
+        friend void to_json(nlohmann::json& j, AnimationTrack const& track);
+        friend void from_json(nlohmann::json const& j, AnimationTrack& track);
 
         std::vector<std::shared_ptr<Keyframe>> m_keyframes;
