+++ include/moth_ui/animation_track_controller.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "moth_ui/animation_track.h"
+#include "moth_ui/ui_fwd.h"
+
+namespace moth_ui {
+    class AnimationTrackController {
+    public:
+        AnimationTrackController(float& target, AnimationTrack& track);
+
+        auto GetTarget() const { return m_track.GetTarget(); }
+        void SetFrame(float frame);
+
+    private:
+        float& m_target;
+        AnimationTrack& m_track;
+    };
+}
