+++ include/moth_ui/layout/layout_entity_group.h
@@ -2,6 +2,7 @@
 
 #include "moth_ui/layout/layout_entity.h"
 #include "moth_ui/animation_clip.h"
+#include "moth_ui/animation_event.h"
 
 #include <nlohmann/json_fwd.hpp>
