+++ src/layout/layout_entity_group.cpp
@@ -1,7 +1,7 @@
 #include "common.h"
 #include "moth_ui/layout/layout_entity_group.h"
-#include "moth_ui/group.h"
 #include "moth_ui/animation_clip.h"
+#include "moth_ui/animation_event.h"
 
 namespace moth_ui {
     LayoutEntityGroup::LayoutEntityGroup(LayoutRect const& initialBounds)
